+++ src/Models/UnixUser.php
@@ -96,6 +96,10 @@ class UnixUser extends ClusterModel implements Model
 
     public function setDefaultNodejsVersion(?string $defaultNodejsVersion): UnixUser
     {
+        Validator::value($defaultNodejsVersion)
+            ->pattern('^[0-9]{1,2}\.[0-9]{1,2}$')
+            ->validate();
+
         $this->defaultNodejsVersion = $defaultNodejsVersion;
 
         return $this;